requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
starlette
passlib
bcrypt
python-multipart
python-jose[cryptography]
